Fix unwind segues
If a modal vc is presented over a sliding view controller then when the modal vc is dismissed the unwind segue returned is a custom segue that doesn’t know how to handle dismissing a modal vc. In this case super should be called if under left or right isn’t the destination.